+++ docs/src/examples/QasBtn/ExWithLoading.vue
@@ -42,12 +42,12 @@ const loading = ref({
 async function handleLoading (fieldKey) {
   loading.value[fieldKey] = true
 
-  await promiseSimulator()
+  await sleepPromise()
 
   loading.value[fieldKey] = false
 }
 
-function promiseSimulator () {
+function sleepPromise () {
   return new Promise(resolve => setTimeout(() => resolve(), 3000))
 }
 </script>