+++ ui/src/components/btn/QasBtn.vue
@@ -9,7 +9,9 @@
       <q-spinner v-if="hasLeftSpinner" :class="iconClasses" size="sm" />
       <q-icon v-if="hasIcon" :class="iconClasses" :name="props.icon" size="xs" />
 
-      <span v-if="showLabel">{{ props.label }}</span>
+      <span v-if="showLabel">
+        {{ props.label }}
+      </span>
 
       <q-spinner v-if="hasRightSpinner" :class="iconRightClasses" size="sm" />
       <q-icon v-if="hasIconRight" :class="iconRightClasses" :name="props.iconRight" size="xs" />