+++ docs/src/examples/QasActionsMenu/ExWithLoading.vue
@@ -1,6 +1,9 @@
 <template>
   <div class="container q-py-lg">
-    <div class="q-mb-sm text-body1">Clique em uma ação para visualizar o exemplo com loading</div>
+    <div class="q-mb-sm text-body1">
+      Clique em uma ação para visualizar o exemplo com loading
+    </div>
+
     <qas-actions-menu v-bind="props" />
 
     <qas-actions-menu v-bind="propsWithoutSplit" />