+++ CHANGELOG.md
@@ -15,10 +15,13 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 - `QasBtn`: modificado a forma de exibição do loading, agora é exibido no lugar do ícone. Poderá haver breaking change visual.
 
 ### Corrigido
-- `QasActionsMenu`: adicionado validação no `q-list` do `qas-actions-menu` pra não exibir um `q-list` vazio quando só há um item.
+- `QasActionsMenu`: 
+  - adicionado validação no `q-list` do `qas-actions-menu` pra não exibir um `q-list` vazio quando só há um item.
+  - corrigido exibição de loading no componente dentro do dropdown ao ter uma prop `loading: true`.
 - `QasHeader`: corrigido exibição do container de descrição em casos de não ter `label` e `description`.
-- `QasTableGenerator`: adicionado `@click.prevent.stop` para não chamar o `rowRouteFn` e `onRowClick` quando se tem ações com botão na tabela.
-- `QasActionsMenu`: corrigido exibição de loading no componente dentro do dropdown ao ter uma prop `loading: true`.
+- `QasTableGenerator`: 
+  - adicionado `@click.prevent.stop` para não chamar o `rowRouteFn` e `onRowClick` quando se tem ações com botão na tabela.
+  - corrigido hover da linha no Safari onde era pego toda tabela ao invés da própria linha.
 
 ### Removido
 - `QasTableGenerator`: removido possibilidade de ter sort na tabela.