+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Neste arquivo (CHANGELOG.MD) você encontrará somente as mudanças referentes a
 ### Sobre os "BREAKING CHANGES"
 Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de serem pequenas, podem alterar o comportamento da funcionalidade caso não seja feita uma atualização, **preste muita atenção** nas breaking changes dentro das versões quando existirem.
 
-## Não publicado
+## [3.18.0-beta.4] - 06-05-2025
 ## BREAKING CHANGE
 - `QasBtn`: modificado a forma de exibição do loading, agora é exibido no lugar do ícone. Poderá haver breaking change visual. Verificar lugares onde é utilizado o `use-ellipsis` também.
 
@@ -3878,3 +3878,4 @@ Adicionado suporte para Pinia/Vuex Seguindo os padrões da biblioteca `@bildvitt
 [3.18.0-beta.1]: https://github.com/bildvitta/asteroid/compare/v3.18.0-beta.0...v3.18.0-beta.1?expand=1
 [3.18.0-beta.2]: https://github.com/bildvitta/asteroid/compare/v3.18.0-beta.1...v3.18.0-beta.2?expand=1
 [3.18.0-beta.3]: https://github.com/bildvitta/asteroid/compare/v3.18.0-beta.2...v3.18.0-beta.3?expand=1
+[3.18.0-beta.4]: https://github.com/bildvitta/asteroid/compare/v3.18.0-beta.3...v3.18.0-beta.4?expand=1